+++ IsoSpecPy/IsoSpecPy/isoFFI.py
@@ -85,7 +85,7 @@ class IsoFFI:
         extern const int elem_table_atomicNo[NUMBER_OF_ISOTOPIC_ENTRIES];
         extern const double elem_table_probability[NUMBER_OF_ISOTOPIC_ENTRIES];
         extern const double elem_table_mass[NUMBER_OF_ISOTOPIC_ENTRIES];
-        extern const int elem_table_massNo[NUMBER_OF_ISOTOPIC_ENTRIES];
+        extern const double elem_table_massNo[NUMBER_OF_ISOTOPIC_ENTRIES];
         extern const int elem_table_extraNeutrons[NUMBER_OF_ISOTOPIC_ENTRIES];
         extern const char* elem_table_element[NUMBER_OF_ISOTOPIC_ENTRIES];
         extern const char* elem_table_symbol[NUMBER_OF_ISOTOPIC_ENTRIES];
